### 2025-06-16 fd34291
trimming paper
