+++ paper/paper.bib
@@ -5,7 +5,8 @@
   volume={113},
   pages={115--129},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.jhin.2021.04.012}
 }
 
 @article{donkin2017,
@@ -15,7 +16,8 @@
   volume={92},
   pages={142--151},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.envsoft.2017.01.020}
 }
 
 @article{ferkol2014,
@@ -26,7 +28,8 @@
   number={3},
   pages={404--406},
   year={2014},
-  publisher={American Thoracic Society}
+  publisher={American Thoracic Society},
+  doi={https://doi.org/10.1513/AnnalsATS.201311-405PS  }
 }
 
 @article{qu2025,
@@ -36,7 +39,8 @@
   volume={10},
   number={10},
   pages={e189151},
-  year={2025}
+  year={2025},
+  doi={10.1172/jci.insight.189151}
 }
 
 @article{levine2022,
@@ -47,7 +51,8 @@
   number={5},
   pages={1153--1154},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.chest.2022.01.014}
 }
 
 @article{masison2021,
@@ -58,7 +63,8 @@
   number={20},
   pages={e2024287118},
   year={2021},
-  publisher={National Academy of Sciences}
+  publisher={National Academy of Sciences},
+  doi={https://doi.org/10.1073/pnas.2024287118}
 }
 
 @article{minucci2020,
@@ -68,7 +74,8 @@
   volume={6},
   pages={36},
   year={2020},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
+  doi={https://doi.org/10.3389/fams.2020.00036}
 }
 
 @article{mitaka2021,
@@ -79,7 +86,8 @@
   number={9},
   pages={993--1001},
   year={2021},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={https://doi.org/10.1111/myc.13292}
 }
 
 @article{neofytos2013,
@@ -90,7 +98,8 @@
   number={3},
   pages={233--242},
   year={2013},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={https://doi.org/10.1111/tid.12060}
 }
 
 @article{pappas2010,
@@ -101,7 +110,8 @@
   number={8},
   pages={1101--1111},
   year={2010},
-  publisher={The University of Chicago Press}
+  publisher={The University of Chicago Press},
+  doi={https://doi.org/10.1086/651262}
 }
 
 @article{ribeiro2022,
@@ -112,7 +122,8 @@
   number={189},
   pages={20210806},
   year={2022},
-  publisher={The Royal Society}
+  publisher={The Royal Society},
+  doi={https://doi.org/10.1098/rsif.2021.0806}
 }
 
 @article{ribeiro2023,
@@ -123,7 +134,8 @@
   number={1},
   pages={6},
   year={2023},
-  publisher={Springer}
+  publisher={Springer},
+  doi={https://doi.org/10.1007/s00285-023-01940-6}
 }
 
 @article{yue2022,
@@ -134,8 +146,7 @@
   number={1},
   pages={37},
   year={2022},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={https://doi.org/10.1038/s41540-022-00247-4}
 }
-
-
 
