+++ paper/paper.md
@@ -1,16 +1,15 @@
 ---
 title: 'The Pulmonary Agent-based Infection simulator (PAI): A Multi-Scale Agent-Based Model of Pulmonary Host-Pathogen Interactions'
 tags:
-  - Agent-Based Model
+  - Agent-Based-Model
   - Infection
   - Immune System
   - Computational Biology
 authors:
-  - name: Henrique de Assis Lopes Ribeiro
+  - name: Henrique AL Ribeiro
     orcid: 0000-0001-9802-5211
     -affiliation: 1 
-    -corresponding: true # (This is how to denote the corresponding author)
-    -email: henrique.deassis@medicine.ufl.edu
+    -corresponding: true
   - name: Sandra A Tsiorintsoa
     -orcid: 0000-0002-7925-987X
     affiliation: 1
@@ -26,74 +25,45 @@ bibliography: paper.bib
 
 # Summary
 
-Understanding the spatial and regulatory dynamics of the immune response to pulmonary infections remains a central challenge in computational immunology. While mechanistic models, such as ordinary differential equations (ODEs) and partial differential equations (PDEs), offer insights into intracellular and population-level processes, they are limited in representing discrete cellular behaviors and spatial heterogeneity. The pulmonary Agent-based Infection simulator (PAI) addresses this gap through an extensible, multi-scale agent-based modeling framework designed to simulate lung immune responses to pathogens such as *Aspergillus fumigatus, Klebsiella pneumoniae*, and SARS-CoV-2.
+Understanding the spatial and regulatory dynamics of the immune response to pulmonary infections remains a central challenge in computational immunology. While mechanistic models, such as ODEs and PDEs, offer insights into intracellular processes, they are limited in representing discrete behaviors and spatial heterogeneity. The pulmonary Agent-based Infection simulator (PAI) addresses this gap through an extensible, multi-scale agent-based modeling designed to simulate lung immune responses to pathogens such as *Aspergillus fumigatus*.
 
-PAI operates within a three-dimensional voxelized space that approximates the alveolar microenvironment, incorporating immune cells (macrophages and neutrophils), structural cells (type II pneumocytes), and fungal agents (conidia and hyphae). It models intercellular signaling via diffusing cytokines and chemokines, immune cell recruitment and movement through chemotactic gradients, and pathogen nutrient acquisition strategies. Intracellular regulation is captured using Boolean Networks that govern phenotype transitions such as activation, cytokine secretion, and apoptosis.
+PAI operates within a three-dimensional voxelized space that approximates the alveolar microenvironment, incorporating host and fungal cells. It models inter- and intracellular signaling, immune cell recruitment and movement, and pathogen nutrient acquisition.
 
-The model is implemented in both Java (jPAI) and C++ (PAI++), with identical outputs and comparable performance, though PAI++ significantly reduces memory usage. With its modular architecture and biological fidelity, PAI provides a valuable tool for simulating complex host-pathogen interactions and for conducting in silico experiments where empirical approaches are limited.
-
+PAI is implemented in both Java (jPAI) and C++ (PAI++), with identical outputs and comparable performance, though PAI++ significantly reduces memory usage. With its modular architecture and biological fidelity, PAI provides a valuable tool for conducting in silico experiments where empirical approaches are limited.
 
 # Statement of Need
 
[…]
