+++ paper/paper.md
@@ -50,8 +50,6 @@ As shown in Figure 1A, at the top level, a Main module orchestrates the initiali
 
 The simulation domain is discretized into a three-dimensional voxel grid, where each voxel represents a microenvironment (~40 µm per side) and maintains localized lists of agents. The Voxel class governs local interactions (interact method), molecular degradation (degrade method), and cell updates (update method) and is invoked iteratively by the Exec class. The Exec class also handles molecular diffusion (via PDE solvers), cellular recruitment (based on chemokine gradients), and agent-level garbage collection.
 
-Agents in the PAI are broadly categorized as Cells or Molecules, both inheriting from an abstract Interactable class (Figure 1B). This base class implements a symmetric interact method based on an “inductive” architecture: new agent types define how they interact with existing agents. The templateInteract method governs the specific pairwise interaction logic, while agent behavior (e.g., migration, phagocytosis, secretion) is defined in subclass implementations. If Agent “A” does not provide code to interact with Agent “B,” nor does Agent “B” provide code to interact with “A,” they learn on the fly that they do not interact (Figure 2).
-
 Agents in the PAI are broadly categorized as Cells or Molecules, both inheriting from an abstract Interactable class (Figure 1B). This base class implements a symmetric interact method based on an “inductive” architecture: new agent types define how they interact with existing agents. If Agent “A” does not provide code to interact with Agent “B,” nor does Agent “B” provide code to interact with “A,” they learn on the fly that they do not interact (Figure 2).
 
 ![Figure 2: Dynamics of interaction between two agents. Decision diagram made by the methods “interact” and “templateInteract” from the two classes trying to interact (“A” and “B”). \label{fig:inductive}](InductiveArchitecture.png)
