### 2025-06-17 9867c7a
add benchmark sentence
