+++ paper/paper.md
@@ -38,7 +38,7 @@ Invasive Pulmonary aspergillosis is a human infection with increasing incidence 
 
 Understanding the complex interplay between immune cells, pathogens, and signaling molecules during pulmonary infections requires computational models that capture both the spatial and regulatory dynamics of the immune response (@minucci2020; @yue2022). While mechanistic models such as ODEs or PDEs are effective at simulating intracellular or population-level behavior, they lack the flexibility to represent discrete cellular interactions and heterogeneous spatial environments.
 
-![Figure 1: ABM Architecture. A: Overall architecture. The “Exec” iterates through the grid-of-voxels (green) and executes the public methods of Voxel (blue). The Voxels iterates through their local lists of agents (orange bottom), executing their public methods. Exec also executes molecule diffusion, “garbage collection (gc)” of dead cells, and cell recruitment. B: Diagram showing the agent’s classes. Black arrows represent inheritance, while red represents composition. All agents are interactables, which means that they can potentially interact with any other agents. The class Interactions provides a library of pairwise agent interaction methods to be called by the agents. \label{fig:architecture}](architecture.png)
+![Figure 1: ABM Architecture. A: Overall architecture. The “Exec” iterates through the grid-of-voxels (green) and executes the public methods of Voxel (blue). The Voxels iterates through their local lists of agents (orange bottom), executing their public methods. Exec also executes molecule diffusion, “garbage collection (gc)” of dead cells, and cell recruitment. B: Diagram showing the agent’s classes. Black arrows represent inheritance, while red represents composition. All agents are interactables, which means that they can potentially interact with any other agents. \label{fig:architecture}](architecture.png)
 
 PAI addresses this gap by providing a scalable, extensible agent-based modeling framework for simulating the tissue-scale immune response to lung infections, such as *Aspergillus fumigatus*. PAI has been used before, and the mathematical and biological components are better explained in @ribeiro2022. While in @ribeiro2023, we modified it to simulate and generate a hypothesis about Covid-Associated-Pulmonary-Aspergillosis. @qu2025 expanded the original model to simulate the effect of NETosis and hemorrhage on Aspergillosis. However, none of these papers provides detailed information about the software. 
 
