### 2025-07-07 #8411 @deassisinfo
Hey, it's been three weeks since I submitted the paper, and I'm worried that there's something I need to do to move the process forward. I'm not complaining, just asking, because if there is something from my part that I don't know, then this paper would be stalled forever. I understand that the note above (regarding the five similar papers) is not intended for me.
