### 2025-11-22 #8411 @majensen
Hello @lgiannantoni - would you have interest in reviewing [this work](#8411) for [JOSS](https://joss.theoj.org)? I could use your expertise -- thanks for considering  --@majensen
