### 2026-01-21 6e7f7b5
Added general README and a Dockerfile
