+++ README.md
@@ -1,0 +1,281 @@
+# PAI
+
+java Pulmonary Agent-based Infection simulator
+
+This project is a C++ and Java cross-platform project. Buidable on Windows, Linux, and macOS via terminal and has a built docker image.
+
+## Using this project with Docker
+
+Without installing any dependencies, this project has a ready-to-run Docker image.
+
+- Docker version 29.1.3 (https://www.docker.com/)
+
+### Pull the Image 
+
+```bash
+docker pull ghcr.io/nutritionallungimmunity/pai/paiimage
+```
+### Run the Container
+
+**For C++**: 
+```bash
+docker run --rm ghcr.io/nutritionallungimmunity/pai/paiimage cpp <num_neutrophils> <num_aspergillus> <num_macrophages> <num_pneumocytes>
+```
+Example
+```bash
+docker run --rm ghcr.io/nutritionallungimmunity/pai/paiimage cpp 15 1920 15 640
+```
+
+**For Java**: 
+```bash
+docker run --rm ghcr.io/nutritionallungimmunity/pai/paiimage java <num_neutrophils> <num_aspergillus> <num_macrophages> <num_pneumocytes>
+```
+Example
+```bash
+docker run --rm ghcr.io/nutritionallungimmunity/pai/paiimage java 15 1920 15 640
+```
+
+**For C++ test**: 
+```bash
+docker run --rm ghcr.io/nutritionallungimmunity/pai/paiimage ctest
+```
+
+**For Java test**: 
+```bash
+docker run --rm ghcr.io/nutritionallungimmunity/pai/paiimage jtest
+```
+
+## To create a local image
+```bash
+docker build -t <myimage> .
+```
+Example
+```bash
+docker build -t paiimage .
+```
+
+Without a Docker, this project can be build and run separately.
+
+# PAIpp
+
+This is the C++ cross-platform project built with CMake. Buidable on Windows, Linux, and macOS via terminal.
+
+## Build Instructions
+
+See [INSTALL.md](INSTALL.md) in **cpp/** directory for platform-specific instructions.
+
+### Quick Start (Unix/macOS/Linux)
+
+```bash
+git clone https://github.com/NutritionalLungImmunity/PAI.git
+cd PAI/cpp
+cmake -S . -B build
+cmake --build build
+./build/PAIpp <num_neutrophils> <num_aspergillus> <num_macrophages> <num_pneumocytes>
+```
+
+### macOS
+
+#### Requirements
+
+- Xcode CLI Tools (for clang++ and make)
+
+```bash
+xcode-select --install
+brew install cmake
+```
+#### Build
+
+```bash
+cmake -S . -B build
+cmake --build build
+./build/PAIpp <num_neutrophils> <num_aspergillus> <num_macrophages> <num_pneumocytes>
+```
+
+### Windows (MinGW)
+
+#### Requirements
+
+- MinGW
+- Add mingw32-make to PATH
+- CMake
+
+#### Build
+
+```bash
+cmake -G "MinGW Makefiles" -S . -B build
+cmake --build build
+./build/PAIpp.exe <num_neutrophils> <num_aspergillus> <num_macrophages> <num_pneumocytes>
+```
+Example
+```bash
+./build/PAIpp.exe 15 1920 15 640
+```
+
+From Command prompt or PowerShell:
+
+```batch
+cmake -G "MinGW Makefiles" -S . -B build
+cmake --build build
+.\build\PAIpp.exe <num_neutrophils> <num_aspergillus> <num_macrophages> <num_pneumocytes>
+```
+Example
+```batch
+.\build\PAIpp.exe 15 1920 15 640
+```
+### Windows (Cygwin)
+
+#### Requirements
+
+- Cygwin
+- Add cygwin to PATH
+- CMake
+
[…]

[…]
