### 2026-01-26 #8411 @majensen
Hi @stsiori - any progress on dockerizing the tool?
