### 2026-01-27 #8411 @stsiori
Hi @majensen - the docker image "pai" is now available under the repository package and the top-level README has all the instructions.  Thank you.
