### 2026-02-07 #8411 @majensen
@stsiori - sorry I'm so delayed - I am getting a 'denied' reponse when I try to pull the image, even when I'm logged in to GitHub. Can you see if you can open it up? Thanks
