### 2026-02-16 #8411 @stsiori
Hello @majensen - sorry for the delay. The image is now public. Thank you.
