### 2026-03-23 #10245 @editorialbot
[REVIEW]: The Pulmonary Agent-based Infection simulator (PAI): A Multi-Scale Agent-Based Model of Pulmonary Host-Pathogen Interactions
**Submitting author:** @deassisinfo (<a href="http://orcid.org/0000-0001-9802-5211">Henrique Assis Lopes Ribeiro</a>)
**Repository:** https://github.com/NutritionalLungImmunity/PAI
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
