### 2026-03-23 #10245 @majensen
> # JOSS Review — PAI: The Pulmonary Agent-based Infection Simulator
> **Reviewer:** @dcnickle
> **Issue:** openjournals/joss-reviews#8411
> **Software version reviewed:** v1.0.0
> **Docker image:** `ghcr.io/nutritionallungimmunity/pai:latest` (digest `sha256:a77fa8b8d74b9354e9ddd552608cc545c4936ff0fbeafd1bb7e462a0031bb72d`)
> **Review platform:** macOS (Apple Silicon, aarch64) with Docker Desktop 29.2.1, `--platform linux/amd64`
>
> ---
>
> ## Conflict of Interest
>
> - [x] I confirm that I have read the JOSS [conflict of interest policy]
>
> ## Code of Conduct
>
> - [x] I confirm that I read and will adhere to the [JOSS code of conduct]
>
> ---
>
> ## General Checks
>
> - [x] **Repository:** Source code is publicly available at https://github.com/NutritionalLungImmunity/PAI.
> - [x] **License:** MIT License, confirmed valid OSI-approved open source license.
> - [x] **Contribution and authorship:** The submitting author (@deassisinfo, 63 commits) has made major contributions. A second author (henriquedeassis, 39 commits) has also contributed substantially. Authorship is consistent with commit history.
> - [x] **Substantial scholarly effort:** The software has been applied in at least three peer-reviewed publications (Ribeiro 2022, Ribeiro 2023, Qu 2025) and implements a multi-scale, multi-language agent-based modeling framework with non-trivial biological scope.
>
> ---
>
> ## Functionality
>
> - [x] **Installation:** The Docker image pulls and runs successfully via `docker pull --platform linux/amd64 ghcr.io/nutritionallungimmunity/pai:latest`. The image is functional.
>
>   > **Required fix:** The image is built only for `linux/amd64`. On Apple Silicon (aarch64), users must explicitly pass `--platform linux/amd64` or the pull will fail with `no matching manifest for linux/arm64/v8`. This must be documented in the README. A multi-arch build (adding `linux/arm64`) would be better.
>
> - [x] **Functionality:** Both the C++ (`PAI++`) and Java (`jPAI`) implementations execute simulations and produce time-series output. The built-in test suites (`ctest`, `jtest`) pass on the first attempt. Key state variables — *Aspergillus fumigatus* burden, TNF-α, and macrophage counts — are verified within expected ranges at six time points (t = 0, 180, 360, 720, 1440, and 2145 minutes) across two simulation runs per implementation.
>
> - [x] **Performance:** `Benchmark.pdf` was provided and reviewed. It contains 8 figures (S1–S8) covering output fidelity, runtime, and memory across 100 simulation runs each.
>
> ---
>
> ## Documentation
>
> - [ ] **Installation instructions:** There are no instructions for users who wish to build from source (C++ compilation flags, JDK version requirements, linking the precompiled C++ modules into the Java layer). Non-Docker users have no clear way to installation.
>
> - [ ] **Example usage:** There is no worked example in the documentation. The `cpp` entrypoint prints usage (`./PAIpp <t0> <Aspergillus> <Macrophage> <Pneumocyte>`) when called without arguments, but no example showing biologically meaningful initial conditions, expected runtime, or how to redirect and interpret the output is provided.
>
> - [ ] **Functionality documentation:** Both implementations write simulation output exclusively to stdout as space-delimited numerical columns with no header row. The C++ implementation produces 20 columns; the Java implementation produces 24. The paper explains that jPAI "contains other pieces of code not discussed here," which accounts for the divergence, but this is not documented anywhere for the user. The paper also states that documentation is provided only for the Java version; this limitation should be stated explicitly in the README.
>
> - [x] **Automated tests:** Both implementations include regression test suites (`ctest` for PAI++, `jtest` for jPAI) that run simulations and assert that state variables remain within biologically expected ranges. Tests pass.
>
> - [ ] **Community guidelines:** No `CONTRIBUTING.md` or equivalent document is present. There are no stated guidelines for reporting bugs, requesting features, or contributing code.
>
> ---
>
> ## Software Paper
>
> - [x] **Summary:** Present. PAI is described as a multi-scale ABM of the pulmonary immune response to *Aspergillus fumigatus*, implemented in both Java and C++, operating on a 3D voxelized alveolar space with intracellular models, molecular diffusion, and immune cell recruitment.
>
> - [x] **Statement of need:** I think this is good. The authors provide epidemiological context (30–60% mortality in invasive aspergillosis, 10–14% incidence in ICU COVID-19 patients) and clearly articulate the limitation of ODE/PDE models in representing discrete cellular interactions and spatial heterogeneity.
>
> - [x] **State of the field:** The paper distinguishes PAI from ODE/PDE-based approaches and positions it relative to existing ABM frameworks, citing prior applications of the model.
>
> - [ ] **Software availability statement:** The paper has no explicit software availability section. The repository URL and version appear in the JOSS submission metadata but are not stated in the paper body.
>
> - [ ] **Quality of writing:** The writing is generally clear.
>   - The paper does not describe the output format, column names, or units, making it impossible for a new user to connect the stdout stream to the described biology without consulting prior publications.
>
>
> - [ ] **References:** Nine references have missing or unverified DOIs as flagged by editorialbot.
>
> ---
>
> ## Bug Reports
>
> The following bugs were identified during functional testing and must be addressed:
>
> **Bug 1 — Java crashes without arguments (Minor)**
> Running `docker run ... java` with no arguments raises an unhandled `ArrayIndexOutOfBoundsException: Index 0 out of bounds for length 0` at `Main.java:38`. The C++ entrypoint handles the missing-arguments case gracefully by printing usage. The Java entrypoint must do the same.
>
> ---
>
> ## Summary Assessment
>
> The software is functional, tests pass, and the biological motivation is sound. The codebase represents a scholarly effort applied to a clinically important problem.

 _Originally posted by @DavidNickle in [#8411](https://github.com/openjournals/joss-reviews/issues/8411#issuecomment-4111183277)_
