+++ cpp/src/edu/uf/main/print/PrintBaseModel.h
@@ -43,6 +43,19 @@ public:
     //Voxel**** grid;
 
 	virtual void printStatistics(int k, const string& file) {
+        if(k == 0) {
+			ostringstream str;
+            str << "Iterations\tTotal_Afumigatus\tResting_Conidia\tSwelling_Conidia\tGerminating_Conidia\tHyphae\t" <<
+				"Total_TAFC\tFree-TAFC\tTAFC_Bound_to_Iron\t" <<
+				"Apolactoferrin\tLactoferrin_Bound_to_Iron\tLactoferrin_Bount_to_two_Iron\t" <<
+				"Total_Transferrin\tApotransferrin\tTransferrin_Bound_to_Iron\tTransferrin_Bount_to_two_Iron\t"  <<
+				"TGF-b\tIL10\tTNF-a\tMIP1-b\tMIP-2\t" <<
+				"Macrophages\tType-II-Pneumocytes\tNeutrophils";
+            
+            string output = str.str();
+            cout << output << endl;
+        }
+
         if (k % 15 != 0) return;
 
         ostringstream str;
