### 2026-03-26 1601d8a
Update test_jpai.py
