### 2026-03-26 e931272
Update PrintBaseModel.java
