### 2026-03-26 #10245 @deassisinfo
We have fixed the bug with the Java version (Java with no parameters), and we have expanded both Java and C++ to show a usage example with a meaningful biological initial value from my Paper (Ribeiro, HAL et al. 2022). We will also indicate that on the README.md shortly. We also added the column names in both Java and C++.

We would also like to point out that both Java and C++ output the same 24 columns. The reviewers must be working with a slightly outdated version of the code. As for the pieces of code in Java that are not discussed, that is true, but those parts are not being executed and should not create any divergence. We only find minimal divergence between Java and C++, which we attributed to implementation differences. But again, those are minimal.

We are working on the other issues the reviewers pointed out, including the Apple Silicon (aarch64) image problem and the revisions needed in the paper, README.md, and other materials.
