### 2026-04-03 86ad108
Create CODE_OF_CONDUCT.md
