### 2026-04-03 80d190e
Create CONTRIBUTING.md
