+++ CONTRIBUTING.md
@@ -1,0 +1,45 @@
+# Contributing to PAI
+
+Thank you for your interest in contributing to the project PAI. We welcome thoughtful, well-prepared contributions that improve the quality, reliability, and usability of the repository.
+
+These are mostly guidelines for smooth contributions, not rules.
+
+
+## Code of Conduct
+PAI project and everyone participating in it is governed by the [PAI code of conduct](CODE_OF_CONDUCT.md).
+
+## Questions
+For any questions, support, or ideas, please don't file an issue but create a GitHub discussion or refer to [GitHub support](https://support.github.com/).
+
+## How to Contribute
+
+
+### Reporting Bugs
+
+To report a bug, open an issue if:
+- The bug is not a **security vulnerability**.
+- The bug was not already reported by searching on GitHub under Issues.
+
+Include a **title and clear description**, as much relevant information as possible (screenshots or logs, if helpful), a **code sample** or an **executable test case** demonstrating the expected behavior that is not occuring. Follow our [bug report template](.github/ISSUE_TEMPLATE/bug_report.md)
+
+If the bug is a **security vulnerability**, please disclose responsibly by letting our team upfront.
+### Pull request
+Your ideas and suggestions are welcome! Please create a new issue using our [custom issue template](.github/ISSUE_TEMPLATE/custom.md) or [feature request template](.github/ISSUE_TEMPLATE/feature_request.md) about any new ideas before investing your time on pull request.
+
+When you are ready to make a [pull request](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request):
+- Fork the repository, create a new branch for your changes, fixes, or features.
+- Update, run tests and check formatting.
+- Push your branch to your fork.
+- Open a new GitHub pull request with the patch you write that fixes a bug or your new changes.
+- Clearly describe the problem and solution in the pull request description with the relevant issue number applicable.
+
+
+## Coding guidelines
+Please follow the existing style of the codebase. Format code consistently, add tests for new behavior, and keep changes focused and small.
+
+## Commit changes
+Write short, clear commit messages that explain what changed. 
+
+## Review process
+Our team will review pull requests as time allows. We may request changes before merging.
+
