### 2026-04-03 8cb89e9
Move INSTALL.md from cpp to the root.
