+++ cpp/INSTALL.md
@@ -1,243 +1,0 @@
-# Installation Instructions
-
-This project uses CMake to configure and build on multiple platforms. Below are the platform-specific instructions.
-
-## Windows (MinGW) 
-
-### Prerequisites
-
-- **MinGW**: Install MinGW to get a compatible GCC toolchain. You can download MinGW from [here](https://sourceforge.net/projects/mingw/). Mark **ming32-gcc-g++** and **mingw32-make** for installation and add **MinGW\bin** to PATH.
-- **CMake**: Download and install CMake from [here](https://cmake.org/download/).
-
-### Building the Project
-
-1. Open a Command Prompt (or PowerShell) and navigate to the project directory.
-2. Create a build directory:
-   
-   ```bash/batch
-   mkdir build
-   cd build
-   ```
-3. Run CMake to configure the project:
-    
-
-    ```bash/batch
-    cmake -G "MinGW Makefiles" ..
-    ```
-
-4. Build the project:
-   
-    ```bash/batch
-    mingw32-make
-    ```
-5. Or combined codes: 
-
-```bash/batch
-    cmake -S -B build -G "MinGW Makefiles"
-    cmake --build build
-```
-### Running the Executable
-Run the executable directly:
-
-```bash
-./build/PAIpp.exe <num_neutrophils> <num_aspergillus> <num_macrophages> <num_pneumocytes>
-```
-
-Example:
-```bash
-./build/PAIpp.exe 15 1920 15 640
-```
-```batch
-.\build\PAIpp.exe <num_neutrophils> <num_aspergillus> <num_macrophages> <num_pneumocytes>
-```
-
-Example:
-```batch
-.\build\PAIpp.exe 15 1920 15 640
-```
-## Windows (Visual Studio)
-
-### Prerequisites
-
-- **Visual Studio 2019 or later**: Install Visual Studio and the necessary C++ build tools.
-
-### Building the Project
-
-1. Open a Developer Command Prompt for Visual Studio.
-2. Navigate to the project directoty and create a build directory:
-
-   ```bash/batch
-   mkdir build
-   cd build
-   ```
-3. Run CMake to configure the project:
-
-    ```bash/batch
-    cmake -G "Visual Studio 16 2019" ..
-    ```
-4. Open the solution in Visual Studio and build the project or 
-
-5. Or combined codes: 
-
-```bash/batch
-    cmake -S -B build -G "Visual Studio 16 2019"
-    cmake --build build
-```
-### Running the Executable
-Run the executable directly:
-
-```bash
-./build/PAIpp.exe <num_neutrophils> <num_aspergillus> <num_macrophages> <num_pneumocytes>
-```
-
-Example:
-```bash
-./build/PAIpp.exe 15 1920 15 640
-```
-```batch
-.\build\PAIpp.exe <num_neutrophils> <num_aspergillus> <num_macrophages> <num_pneumocytes>
-```
-
-Example:
-```batch
-.\build\PAIpp.exe 15 1920 15 640
-```
-## Windows (Cygwin)
-
-### Prerequisites
-
-- **Cygwin**: Install MinGW to get a compatible GCC toolchain. You can download MinGW from [here](https://www.cygwin.com/).
-- **CMake**: Download and install CMake from [here](https://cmake.org/download/).
-
-### Building the Project
-
-1. Open a Command Prompt (or PowerShell) and navigate to the project directory.
-2. Create a build directory:
-   
-   ```bash/batch
-   mkdir build
-   cd build
-   ```
-3. Run CMake to configure the project:
-    
-    ```bash/batch
-    cmake -G "Unix Makefiles" ..
-    ```
-4. Build the project:
-    
[…]
