### 2026-04-03 1dcf004
Move INSTALL.md from cpp to the root.
