+++ paper/paper.bib
@@ -29,7 +29,7 @@
   pages={404--406},
   year={2014},
   publisher={American Thoracic Society},
-  doi={https://doi.org/10.1513/AnnalsATS.201311-405PS  }
+  doi={https://doi.org/10.1513/AnnalsATS.201311-405PS}
 }
 
 @article{qu2025,
@@ -148,5 +148,53 @@
   year={2022},
   publisher={Nature Publishing Group UK London},
   doi={https://doi.org/10.1038/s41540-022-00247-4}
+}
+
+@article{celada1992,
+  title={A computer model of cellular interactions in the immune system},
+  author={Celada, Franco and Seiden, Philip E},
+  journal={Immunology today},
+  volume={13},
+  number={2},
+  pages={56--62},
+  year={1992},
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/0167-5699(92)90135-T}
+}
+
+@article{rapin2010,
+  title={Computational immunology meets bioinformatics: the use of prediction tools for molecular binding in the simulation of the immune system},
+  author={Rapin, Nicolas and Lund, Ole and Bernaschi, Massimo and Castiglione, Filippo},
+  journal={PloS one},
+  volume={5},
+  number={4},
+  pages={e9862},
+  year={2010},
+  publisher={Public Library of Science San Francisco, USA},
+  doi={https://doi.org/10.1371/journal.pone.0009862}
+}
+
+@article{basak2021,
+  title={In silico designing of vaccine candidate against Clostridium difficile},
+  author={Basak, Srijita and Deb, Debashrito and Narsaria, Utkarsh and Kar, Tamalika and Castiglione, Filippo and Sanyal, Indraneel and Bade, Pratap D and Srivastava, Anurag P},
+  journal={Scientific Reports},
+  volume={11},
+  number={1},
+  pages={14215},
+  year={2021},
+  publisher={Nature Publishing Group UK London},
+  doi={https://doi.org/10.1038/s41598-021-93305-6}
+}
+
+@article{segovia2004,
+  title={Identifying control mechanisms of granuloma formation during M. tuberculosis infection using an agent-based model},
+  author={Segovia-Juarez, Jose L and Ganguli, Suman and Kirschner, Denise},
+  journal={Journal of theoretical biology},
+  volume={231},
+  number={3},
+  pages={357--376},
+  year={2004},
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.jtbi.2004.06.031}
 }
 
