### 2026-04-06 247157c
State of the Field, Res Impact, Design, and AI Usage added
