### 2026-04-06 #10245 @deassisinfo
I made the necessary changes to the text today (added the State of the Field, Research Statement, and AI Usage Disclosure). As for Software Design, the paper had a section already; it just had a different name, so I renamed the Model section to Software Design. I apologize for the confusion, but we use this wording when writing papers in Computer Biology journals. I've seen that @majensen generated a PDF two days ago. I apologize, but I only updated the text today. There was an issue raised in relation to the DOIs; however, I can not see the problem by myself. When I generate the PDF, everything is fine. All the references are there and have the DOIs. I would like clarification on what exactly the issue is.
Regarding the other issues (not related to the text), the other members of the team and I are looking into them and are almost done.
