+++ paper/paper.bib
@@ -6,7 +6,7 @@
   pages={115--129},
   year={2021},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.jhin.2021.04.012}
+  doi={10.1016/j.jhin.2021.04.012}
 }
 
 @article{donkin2017,
@@ -17,7 +17,7 @@
   pages={142--151},
   year={2017},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.envsoft.2017.01.020}
+  doi={10.1016/j.envsoft.2017.01.020}
 }
 
 @article{ferkol2014,
@@ -29,7 +29,7 @@
   pages={404--406},
   year={2014},
   publisher={American Thoracic Society},
-  doi={https://doi.org/10.1513/AnnalsATS.201311-405PS}
+  doi={10.1513/AnnalsATS.201311-405PS}
 }
 
 @article{qu2025,
@@ -64,7 +64,7 @@
   pages={e2024287118},
   year={2021},
   publisher={National Academy of Sciences},
-  doi={https://doi.org/10.1073/pnas.2024287118}
+  doi={10.1073/pnas.2024287118}
 }
 
 @article{minucci2020,
@@ -75,7 +75,7 @@
   pages={36},
   year={2020},
   publisher={Frontiers Media SA},
-  doi={https://doi.org/10.3389/fams.2020.00036}
+  doi={10.3389/fams.2020.00036}
 }
 
 @article{mitaka2021,
@@ -87,7 +87,7 @@
   pages={993--1001},
   year={2021},
   publisher={Wiley Online Library},
-  doi={https://doi.org/10.1111/myc.13292}
+  doi={10.1111/myc.13292}
 }
 
 @article{neofytos2013,
@@ -99,7 +99,7 @@
   pages={233--242},
   year={2013},
   publisher={Wiley Online Library},
-  doi={https://doi.org/10.1111/tid.12060}
+  doi={10.1111/tid.12060}
 }
 
 @article{pappas2010,
@@ -111,7 +111,7 @@
   pages={1101--1111},
   year={2010},
   publisher={The University of Chicago Press},
-  doi={https://doi.org/10.1086/651262}
+  doi={10.1086/651262}
 }
 
 @article{ribeiro2022,
@@ -123,7 +123,7 @@
   pages={20210806},
   year={2022},
   publisher={The Royal Society},
-  doi={https://doi.org/10.1098/rsif.2021.0806}
+  doi={10.1098/rsif.2021.0806}
 }
 
 @article{ribeiro2023,
@@ -135,7 +135,7 @@
   pages={6},
   year={2023},
   publisher={Springer},
-  doi={https://doi.org/10.1007/s00285-023-01940-6}
+  doi={10.1007/s00285-023-01940-6}
 }
 
 @article{yue2022,
@@ -147,7 +147,7 @@
   pages={37},
   year={2022},
   publisher={Nature Publishing Group UK London},
-  doi={https://doi.org/10.1038/s41540-022-00247-4}
+  doi={10.1038/s41540-022-00247-4}
 }
 
 @article{celada1992,
@@ -159,7 +159,7 @@
   pages={56--62},
   year={1992},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/0167-5699(92)90135-T}
+  doi={10.1016/0167-5699(92)90135-T}
 }
 
 @article{rapin2010,
@@ -171,7 +171,7 @@
   pages={e9862},
   year={2010},
   publisher={Public Library of Science San Francisco, USA},
-  doi={https://doi.org/10.1371/journal.pone.0009862}
+  doi={10.1371/journal.pone.0009862}
 }
 
[…]
