### 2026-04-06 #10245 @deassisinfo
I think I got what the DOI's problem was. Now that I saw it, it's pretty obvious. I believe I have solved it.
