### 2026-05-10 #10245 @majensen
@deassisinfo, can you create a Zenodo archive of the repository for the submission? You can follow [this example](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729). Please make the name of the archive the same as the paper name, and the archive authors the same as the paper authors. Just report back here the DOI of the archive. Thanks!
