### 2026-05-13 3227b2d
Update paper.md
Output Description and Code Availability
