### 2026-05-13 8b3c437
Update README.md
lack of C++ documentation
