+++ paper/paper.md
@@ -38,7 +38,7 @@ Invasive Pulmonary aspergillosis is a human infection with increasing incidence 
 
 Understanding the complex interplay between immune cells, pathogens, and signaling molecules during pulmonary infections requires computational models that capture both the spatial and regulatory dynamics of the immune response (@minucci2020; @yue2022). While mechanistic models such as ODEs or PDEs are effective at simulating intracellular or population-level behavior, they lack the flexibility to represent discrete cellular interactions and heterogeneous spatial environments.
 
-![Figure 1: ABM Architecture. A: Overall architecture. The “Exec” iterates through the grid-of-voxels (green) and executes the public methods of Voxel (blue). The Voxels iterates through their local lists of agents (orange bottom), executing their public methods. Exec also executes molecule diffusion, “garbage collection (gc)” of dead cells, and cell recruitment. B: Diagram showing the agent’s classes. Black arrows represent inheritance, while red represents composition. All agents are interactables, which means that they can potentially interact with any other agents. \label{fig:architecture}](architecture.png)
+![ABM Architecture. A: Overall architecture. The “Exec” iterates through the grid-of-voxels (green) and executes the public methods of Voxel (blue). The Voxels iterates through their local lists of agents (orange bottom), executing their public methods. Exec also executes molecule diffusion, “garbage collection (gc)” of dead cells, and cell recruitment. B: Diagram showing the agent’s classes. Black arrows represent inheritance, while red represents composition. All agents are interactables, which means that they can potentially interact with any other agents. \label{fig:architecture}](architecture.png)
 
 PAI addresses this gap by providing a scalable, extensible agent-based modeling framework for simulating the tissue-scale immune response to lung infections, such as *Aspergillus fumigatus*. PAI has been used before, and the mathematical and biological components are better explained in @ribeiro2022. While in @ribeiro2023, we modified it to simulate and generate a hypothesis about Covid-Associated-Pulmonary-Aspergillosis. @qu2025 expanded the original model to simulate the effect of NETosis and hemorrhage on Aspergillosis. However, none of these papers provides detailed information about the software. 
 
@@ -65,13 +65,13 @@ The simulation domain is discretized into a three-dimensional voxel grid, where 
 
 Agents in the PAI are broadly categorized as Cells or Molecules, both inheriting from an abstract Interactable class (Figure 1B). This base class implements a symmetric interact method based on an “inductive” architecture: new agent types define how they interact with existing agents. If Agent “A” does not provide code to interact with Agent “B,” nor does Agent “B” provide code to interact with “A,” they learn on the fly that they do not interact (Figure 2).
 
-![Figure 2: Dynamics of interaction between two agents. Decision diagram made by the methods “interact” and “templateInteract” from the two classes trying to interact (“A” and “B”). \label{fig:inductive}](InductiveArchitecture.png)
+![Dynamics of interaction between two agents. Decision diagram made by the methods “interact” and “templateInteract” from the two classes trying to interact (“A” and “B”). \label{fig:inductive}](InductiveArchitecture.png)
 
 ## Instantiation
 
[…]
