### 2026-05-25 #10245 @deassisinfo
Hello @majensen , when I downloaded the preview, I found a small problem. The legends had the word Figure repeated, like "Figure 1: Figure 1:" It seems to me that this was caused by the fact that I added the word Figure as part of the legend itself. I went to my repository and removed it, but the problem persisted. However, when I preview the PDF on the git repository, I do not see this problem (even in older versions). Other than that, all is fine.
