### 2026-06-01 9fa30e9
Update paper.md (spelled out USA)
